service/requirements.txt
librosa==0.10.0.post2
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
pytube==12.1.3
PyYAML==6.0
scikit_learn==1.2.2
streamlit==1.20.0
torch==2.0.0
torchaudio==2.0.1
tqdm==4.65.0
